requirements/base.txt
Django==4.1.7
django-phonenumber-field==7.0.2
django-ckeditor==6.5.1
django-modeltranslation==0.18.9
Pillow==9.4.0
django-hitcount==1.3.5
django-rosetta==0.9.8